refactor(schedule): separate parsing, occurrence, and conflict logic
Move schedule parsing, occurrence compilation, conflict checks, and colocated tests into dedicated submodules.

Keep schedule.rs focused on shared types and the public scheduling API.
